// fe-web/src/_generated.ts
export type EnumsRole = 'customer' | 'contractor';